Channel class scope exceptions

+++ src/Channel.php
@@ -24,6 +24,10 @@ class Channel extends BaseMethod
                 throw new ChannelException("Please provide a channel name, or use Twitch::setAccessToken() to define the channels access token.");
             }
 
+            if (!Twitch::isAuthorizedFor('channel_read')) {
+                throw new TwitchScopeException("You do not have sufficient scope priviledges to run this command. Make sure you're authorized for `channel_read`.", 401);
+            }
+
             $this->setEndpoint("channel");
         } else {
             $this->setEndpoint("channels/{$channel_name}");
@@ -48,6 +52,10 @@ class Channel extends BaseMethod
     // to be used with setTitle and setGame
     public function update()
     {
+        if (!Twitch::isAuthorizedFor('channel_editor'))    {
+            throw new TwitchScopeException("You do not have sufficient scope priviledges to run this command. Make sure you're authorized for `channel_editor`.", 401);
+        }
+
         $response = Twitch::api($this->endpoint())->put([
           'channel' => [
               'game' => $this->data()->game,
@@ -58,9 +66,12 @@ class Channel extends BaseMethod
         return $this;
     }
 
-    // the other endpoint options available
     public function resetStreamKey()
     {
+        if (!Twitch::isAuthorizedFor('channel_stream'))    {
+            throw new TwitchScopeException("You do not have sufficient scope priviledges to run this command. Make sure you're authorized for `channel_stream`.", 401);
+        }
+
         return Twitch::api($this->endpoint())->delete();
     }
 
@@ -88,6 +99,10 @@ class Channel extends BaseMethod
 
     public function editors()
     {
+        if (!Twitch::isAuthorizedFor('channel_read')) {
+            throw new TwitchScopeException("You do not have sufficient scope priviledges to run this command. Make sure you're authorized for `channel_read`.", 401);
+        }
+
         $response = Twitch::api($this->endpoint() . "/editors")->get()->data();
 
         if ($response) {
@@ -110,6 +125,10 @@ class Channel extends BaseMethod
 
     public function commercial($length = 30)
     {
+        if (!Twitch::isAuthorizedFor('channel_commercial'))    {
+            throw new TwitchScopeException("You do not have sufficient scope priviledges to run this command. Make sure you're authorized for `channel_commercial`.", 401);
+        }
+
         if (empty($this->data()->partner)) {
             throw new ChannelException("You need to be a partner to run commercials.");
         }
